Delete payment-method-store.ts from wpcom-checkout
StoreStateValue and StoreState were the last two remaining exports.
They are only used by the credit card payment method, so inlined them
into credit-card/types.ts instead. The jetpack-cloud and a8c-for-agencies
credit-card-fields components that used StoreState<string> as an
explicit type annotation now let TypeScript infer it from the selector.
