+++ client/a8c-for-agencies/sections/purchases/payment-methods/payment-method-add/credit-card-fields/index.tsx
@@ -11,7 +11,6 @@ import useStoredCards from '../../hooks/use-stored-cards';
 import CreditCardElementField from './credit-card-element-field';
 import CreditCardLoading from './credit-card-loading';
 import SetAsPrimaryPaymentMethod from './set-as-primary-payment-method';
-import type { StoreState } from '@automattic/wpcom-checkout';
 import type { StripeElementChangeEvent, StripeElementStyle } from '@stripe/stripe-js';
 
 import './style.scss';
@@ -20,10 +19,7 @@ export default function CreditCardFields() {
 	const translate = useTranslate();
 
 	const [ isStripeFullyLoaded, setIsStripeFullyLoaded ] = useState( false );
-	const fields: StoreState< string > = useSelect(
-		( select ) => select( creditCardStore ).getFields(),
-		[]
-	);
+	const fields = useSelect( ( select ) => select( creditCardStore ).getFields(), [] );
 	const useAsPrimaryPaymentMethod: boolean = useSelect(
 		( select ) => select( creditCardStore ).useAsPrimaryPaymentMethod(),
 		[]
