--- packages/wpcom-checkout/src/payment-method-store.ts
@@ -1,7 +1,0 @@
-export interface StoreStateValue {
-	value: string;
-	isTouched: boolean;
-	errors?: string[];
-}
-
-export type StoreState< N extends string > = Record< N, StoreStateValue >;
